Moved convert to code to its own file. also made htmlpreview a single step.
